+++ src/connection/database/login_pane.jsx
@@ -30,6 +30,7 @@ export default class LoginPane extends React.Component {
     const headerText = instructions || null;
     const header = headerText && <p>{headerText}</p>;
 
+    // Should never validate format on login because of custom db connection and import mode
     const fieldPane = usernameStyle === "email"
       ? <EmailPane
           i18n={i18n}
@@ -41,7 +42,7 @@ export default class LoginPane extends React.Component {
           lock={lock}
           placeholder={usernameInputPlaceholder}
           usernameStyle={usernameStyle}
-          validateFormat={true}
+          validateFormat={false}
         />
 
     const passwordPane = showPassword