+++ src/core/index.js
@@ -22,12 +22,14 @@ const {
 export function setup(id, clientID, domain, options, hookRunner, emitEventFn) {
   let m = init(id, Immutable.fromJS({
     clientBaseUrl: extractClientBaseUrlOption(options, domain),
+    tenantBaseUrl: extractTenantBaseUrlOption(options, domain),
     languageBaseUrl: extractLanguageBaseUrlOption(options, domain),
     auth: extractAuthOptions(options),
     clientID: clientID,
     domain: domain,
     emitEventFn: emitEventFn,
     hookRunner: hookRunner,
+    useTenantInfo: options.__useTenantInfo || false,
     allowedConnections: Immutable.fromJS(options.allowedConnections || []),
     ui: extractUIOptions(id, options)
   }));
@@ -53,9 +55,18 @@ export function clientBaseUrl(m) {
   return get(m, "clientBaseUrl");
 }
 
+export function tenantBaseUrl(m) {
+  return get(m, "tenantBaseUrl");
+}
+
+export function useTenantInfo(m) {
+  return get(m, "useTenantInfo");
+}
+
 export function languageBaseUrl(m) {
   return get(m, "languageBaseUrl");
 }
+
 export function setSubmitting(m, value, error = "") {
   m = tset(m, "submitting", value);
   m = clearGlobalSuccess(m);
@@ -102,7 +113,7 @@ export function stopRendering(m) {
 function extractUIOptions(id, options) {
   const closable = options.container ? false : undefined === options.closable ? true : !!options.closable;
   const theme = options.theme || {};
-  const { labeledSubmitButton, logo, primaryColor } = theme;
+  const { labeledSubmitButton, logo, primaryColor, authButtons } = theme;
 
   const avatar = options.avatar !== null;
   const customAvatarProvider = options.avatar
@@ -127,11 +138,24 @@ function extractUIOptions(id, options) {
     mobile: undefined === options.mobile ? false : !!options.mobile,
     popupOptions: undefined === options.popupOptions ? {} : options.popupOptions,
     primaryColor: typeof primaryColor === "string" ? primaryColor : undefined,
-    rememberLastLogin: undefined === options.rememberLastLogin ? true : !!options.rememberLastLogin
+    rememberLastLogin: undefined === options.rememberLastLogin ? true : !!options.rememberLastLogin,
+    authButtonsTheme: typeof authButtons === "object" ? authButtons : {}
   });
 }
 
-const { get: getUIAttribute } = dataFns(["core", "ui"]);
+const { 
+  get: getUI, 
+  set: setUI
+} = dataFns(["core", "ui"]);
+
+const { 
+  get: tgetUI, 
+  set: tsetUI
+} = dataFns(["core", "transient", "ui"]);
+
+const getUIAttribute = (m, attribute) => {
+  return tgetUI(m, attribute) || getUI(m, attribute);
+};
 
 export const ui = {
   containerID: lock => getUIAttribute(lock, "containerID"),
@@ -149,6 +173,7 @@ export const ui = {
   mobile: lock => getUIAttribute(lock, "mobile"),
   popupOptions: lock => getUIAttribute(lock, "popupOptions"),
   primaryColor: lock => getUIAttribute(lock, "primaryColor"),
+  authButtonsTheme: lock => getUIAttribute(lock, "authButtonsTheme"),
   rememberLastLogin: m => tget(
     m,
     "rememberLastLogin",
@@ -159,6 +184,7 @@ export const ui = {
 const { get: getAuthAttribute } = dataFns(["core", "auth"]);
 
 export const auth = {
+  connectionScopes: m => getAuthAttribute(m, "connectionScopes"),
   params: m => tget(m, "authParams") || getAuthAttribute(m, "params"),
   redirect: lock => getAuthAttribute(lock, "redirect"),
   redirectUrl: lock => getAuthAttribute(lock, "redirectUrl"),
@@ -169,6 +195,7 @@ export const auth = {
 
 function extractAuthOptions(options) {
   let {
+    connectionScopes,
     params,
     redirect,
     redirectUrl,
@@ -177,6 +204,7 @@ function extractAuthOptions(options) {
     sso
   } = options.auth || {};
 
+  connectionScopes = typeof connectionScopes === "object" ? connectionScopes : {};
   params = typeof params === "object" ? params : {};
   redirectUrl = typeof redirectUrl === "string" && redirectUrl ? redirectUrl : undefined;
   redirect = typeof redirect === "boolean" ? redirect : true;
@@ -190,6 +218,7 @@ function extractAuthOptions(options) {
   }
 
   return Immutable.fromJS({
+    connectionScopes,
     params,
     redirect,
     redirectUrl,
@@ -210,6 +239,10 @@ function extractClientBaseUrlOption(opts, domain) {
     return opts.clientBaseUrl;
   }
 
+  if (opts.configurationBaseUrl && typeof opts.configurationBaseUrl === "string") {
+    return opts.configurationBaseUrl;
+  }
+
   if (opts.assetsUrl && typeof opts.assetsUrl === "string") {
     return opts.assetsUrl;
   }
@@ -223,6 +256,32 @@ function extractClientBaseUrlOption(opts, domain) {
     return parts.length > 3
       ? "https://cdn." + parts[parts.length - 3] + DOT_AUTH0_DOT_COM
       : AUTH0_US_CDN_URL;
+  } else {
+    return domainUrl;
+  }
+}
+
+function extractTenantBaseUrlOption(opts, domain) {
+  if (opts.configurationBaseUrl && typeof opts.configurationBaseUrl === "string") {
+    return opts.configurationBaseUrl;
+  }
+
+  if (opts.assetsUrl && typeof opts.assetsUrl === "string") {
+    return opts.assetsUrl;
+  }
+
+  const domainUrl = "https://" + domain;
+  const hostname = parseUrl(domainUrl).hostname;
+  const DOT_AUTH0_DOT_COM = ".auth0.com";
+  const AUTH0_US_CDN_URL = "https://cdn.auth0.com";
+  if (endsWith(hostname, DOT_AUTH0_DOT_COM)) {
+    const parts = hostname.split(".");
+    const domain = parts.length > 3
+      ? "https://cdn." + parts[parts.length - 3] + DOT_AUTH0_DOT_COM
+      : AUTH0_US_CDN_URL;
+    const tenant_name = parts[0];
+
+    return `${domain}/tenants/v1/${tenant_name}`
   } else {
     return domainUrl;
   }
@@ -381,6 +440,14 @@ export function loginErrorMessage(m, error, type) {
     code = INVALID_MAP[type];
   }
 
+  if (code === "a0.mfa_registration_required") {
+    code = "lock.mfa_registration_required";
+  }
+
+  if (code === "a0.mfa_invalid_code") {
+    code = "lock.mfa_invalid_code";
+  }
+
   return i18n.str(m, ["error", "login", code])
     || i18n.str(m, ["error", "login", "lock.fallback"]);
 }
@@ -410,6 +477,10 @@ export function emitAuthorizationErrorEvent(m, error) {
   emitEvent(m, "authorization_error", error);
 }
 
+export function emitUnrecoverableErrorEvent(m, error) {
+  emitEvent(m, "unrecoverable_error", error);
+}
+
 export function showBadge(m) {
   return hasFreeSubscription(m) || false;
 }
@@ -421,8 +492,32 @@ export function overrideOptions(m, opts) {
     m = tset(m, "allowedConnections", Immutable.fromJS(opts.allowedConnections));
   }
 
+  if (opts.flashMessage) {
+    const key = "success" === opts.flashMessage.type ? "globalSuccess" : "globalError";
+    m = tset(m, key, opts.flashMessage.text);
+  }
+
   if (opts.auth && opts.auth.params) {
     m = tset(m, "authParams", Immutable.fromJS(opts.auth.params));
+  }
+
+  if (opts.theme) {
+    if (opts.theme.primaryColor) {
+      m = tset(m, ["ui", "primaryColor"], opts.theme.primaryColor);
+    }
+
+    if (opts.theme.logo) {
+      m = tset(m, ["ui", "logo"], opts.theme.logo);
+    }
+  }
+  
+  if (opts.language) {
+    opts.dict = opts.dict || {};
+
+    m = tset(m, ["ui", "language"], opts.language);
+    m = tset(m, ["ui", "dict"], opts.dict);
+    
+    m = i18n.initI18n(m);
   }
 
   if (typeof opts.rememberLastLogin === "boolean") {