+++ src/field/username.js
@@ -8,7 +8,10 @@ const DEFAULT_CONNECTION_VALIDATION = { username: { min: 1, max: 15 } };
 const regExp = /^[a-zA-Z0-9_]+$/;
 
 function validateUsername(str, validateFormat, settings = DEFAULT_CONNECTION_VALIDATION.username) {
-  if (!validateFormat) {
+
+// If the connection does not have validation settings, it should only check if the field is empty.
+// validateFormat overrides this logic to disable validation on login (login should never validate format)
+  if (!validateFormat  || settings == null) {
     return trim(str).length > 0;
   }
 
@@ -30,7 +33,8 @@ function validateUsername(str, validateFormat, settings = DEFAULT_CONNECTION_VAL
 }
 
 export function getUsernameValidation(m) {
-  return databaseConnection(m).getIn(['validation', 'username']).toJS()
+  const usernameValidation = databaseConnection(m).getIn(['validation', 'username']);
+  return usernameValidation ? usernameValidation.toJS() : null;
 }
 
 export function setUsername(m, str, usernameStyle = "username", validateUsernameFormat = true) {